Extract functions to utils file
